+++ buildSrc/src/main/kotlin/SamplesDependencies.kt
@@ -1,8 +1,8 @@
 object SamplesVersions {
     val kotlin = PluginVersions.kotlin
     val androidPlugin = PluginVersions.androidPlugin
-    val gradleWrapper = PluginVersions.gradleWrapper
-    val ktlintCore = "0.29.0"
+    val gradleWrapper = "5.1.1"
+    val ktlintCore = "0.32.0"
     val junit = "4.12"
 
     // Android sample dependencies