+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -1,10 +1,10 @@
 object PluginVersions {
-    val kotlin = "1.3.21"
-    val ktlintPlugin = "7.2.0"
+    val kotlin = "1.3.31"
+    val ktlintPlugin = "7.4.0"
     val gradlePublishPlugin = "0.10.1"
     val androidPlugin = "3.3.0"
     val semver = "1.1.1"
-    val gradleWrapper = "4.10.2"
+    val gradleWrapper = "5.4"
     val junit5 = "5.4.0"
     val assertJ = "3.11.1"
 }