+++ samples/kotlin-rulesets-creating/src/main/kotlin/org/jlleitschuh/gradle/ktlint/sample/kotlin/CustomRuleSetProvider.kt
@@ -1,7 +1,7 @@
 package org.jlleitschuh.gradle.ktlint.sample.kotlin
 
-import com.github.shyiko.ktlint.core.RuleSet
-import com.github.shyiko.ktlint.core.RuleSetProvider
+import com.pinterest.ktlint.core.RuleSet
+import com.pinterest.ktlint.core.RuleSetProvider
 
 class CustomRuleSetProvider : RuleSetProvider {