+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintBasePlugin.kt
@@ -6,7 +6,6 @@ import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.file.ConfigurableFileTree
 import org.gradle.api.tasks.util.PatternFilterable
-import org.gradle.util.GradleVersion
 
 internal typealias FilterApplier = (Action<PatternFilterable>) -> Unit
 internal typealias KotlinScriptAdditionalPathApplier = (ConfigurableFileTree) -> Unit
@@ -19,17 +18,12 @@ open class KtlintBasePlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
         val filterTargetApplier: FilterApplier = {
-            if (GradleVersion.current() >= GradleVersion.version("4.9")) {
-                // API only added in Gradle 4.9
-                target.tasks.withType(KtlintCheckTask::class.java).configureEach(it)
-            } else {
-                target.tasks.withType(KtlintCheckTask::class.java, it)
-            }
+            target.tasks.withType(BaseKtlintCheckTask::class.java).configureEach(it)
         }
 
         val kotlinScriptAdditionalPathApplier: KotlinScriptAdditionalPathApplier = { additionalFileTree ->
             val configureAction = Action<Task> { task ->
-                with(task as KtlintCheckTask) {
+                with(task as BaseKtlintCheckTask) {
                     source = source.plus(additionalFileTree.also {
                         it.include("*.kts")
                     })