+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintExtension.kt
@@ -25,7 +25,7 @@ internal constructor(
     /**
      * The version of ktlint to use.
      */
-    val version: Property<String> = objectFactory.property { set("0.32.0") }
+    val version: Property<String> = objectFactory.property { set("0.33.0") }
 
     /**
      * Enable verbose mode.