+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   - Support for new JS plugin: "org.jetbrains.kotlin.js" (#252)
   - Support for android "com.android.dynamic-feature" plugin (#260)
 ### Changed
-  - Update Kotlin to `1.3.40` version
+  - Update Kotlin to `1.3.41` version
 ### Removed
   - ?
 ### Fixed