+++ CHANGELOG.md
@@ -5,14 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [8.2.0] - 2019-07-18
 ### Added
   - Support for new JS plugin: "org.jetbrains.kotlin.js" (#252)
   - Support for android "com.android.dynamic-feature" plugin (#260)
 ### Changed
   - Update Kotlin to `1.3.41` version
-### Removed
-  - ?
 ### Fixed
   - Usage of bashisms in git hook script (#251)